refactor(svelte2tsx): オープナー後置のディレクティブ文を attributes/directive_suffix.rs へ切り出す
build_element_directive_suffix_segments / build_directive_prefix_suffix /
build_component_directive_suffix を純粋移動。出力・ロジックの変更なし。
