+++ crates/rsvelte_core/src/svelte2tsx/template/attributes/mod.rs
@@ -3,6 +3,7 @@
 pub(super) mod action;
 pub(super) mod binding;
 pub(super) mod class_style;
+pub(super) mod directive_suffix;
 pub(super) mod event_handler;
 pub(super) mod let_;
 pub(super) mod spread;
